refactor: met en valeur les tests api dans le code infra (000)
